requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
django-filter==23.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
Django==4.2
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
djoser==2.2.0
idna==3.4
jwt==1.3.1
oauthlib==3.2.2
Pillow==9.5.0
pip==23.1.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
requests==2.28.2
setuptools==67.7.2
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.4
urllib3==1.26.15
drf_yasg==1.21.5
django-stubs==4.2.4